Fix relationship to check that a level has actually been awarded (fixes #45)
Previously, the algorithm just checked for the existence. However, if you reject a proposal
then the awarded skill entry is still created but with a level of '0'. Rendering this value then
doesn't return any skill string because 0 isn't a valid level. This was also creating incorrect
readings for the total number of awarded skills in a category.

This fixes the issue by also checking that the level is greater than 0 - i.e., it's been awarded.